add state for attempts and note put log to server when user makes 2 moves
