+++ app/src/state/track.ts
@@ -1,0 +1,7 @@
+import { atom } from 'recoil';
+import { TrackAttemptStatus } from '../structs/track';
+
+export const attemptStatus = atom({
+  key: 'attemptsStatus',
+  default: TrackAttemptStatus.INACTIVE,
+});
